Improvements for the connection extensions to safely write data
* Renamed ThreadSafeWrite to ThreadSafeWriter
* Moved the call of PipeWriter.Advance into a separate Commit method which needs to be called at the end of the using statement. This way, it's only sent when all the previous code didn't fail by an exception.
* Added a packet type to the constructor, which now also automatically sets the size of the packet. Since most packet sizes are known up-front this simplifies the most code.
