+++ src/Network/ConnectionExtensions.cs
@@ -30,11 +30,12 @@ namespace MUnique.OpenMU.Network
         /// Starts a safe write to this connection. The specified packet size has to be exact, because it can't be changed afterwards.
         /// </summary>
         /// <param name="connection">The connection.</param>
-        /// <param name="exactPacketSize">Size of the exact packet.</param>
-        /// <returns>An <see cref="IDisposable"/> which can be used in a using-construct. It provides the target span to write at, too.</returns>
-        public static ThreadSafeWrite SafeWrite(this IConnection connection, int exactPacketSize)
+        /// <param name="packetType">Type of the packet (e.g. 0xC1).</param>
+        /// <param name="expectedPacketSize">Size of the expected packet. It's set in the header by default.</param>
+        /// <returns>An <see cref="IDisposable" /> which can be used in a using-construct. It provides the target span to write at, too.</returns>
+        public static ThreadSafeWriter StartSafeWrite(this IConnection connection, byte packetType, int expectedPacketSize)
         {
-            return new ThreadSafeWrite(connection, exactPacketSize);
+            return new ThreadSafeWriter(connection, packetType, expectedPacketSize);
         }
     }
 }
